+++ CHANGELOG.md
@@ -5,11 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
 ### Fixed
 
 - Restored capability to run tests out-of-tree.
-
-## Unreleased
 
 ## [v0.10.0](https://github.com/allenai/tango/releases/tag/v0.10.0) - 2022-07-07