+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Restored capability to run tests out-of-tree.
+
 ## [v0.10.0](https://github.com/allenai/tango/releases/tag/v0.10.0) - 2022-07-07
 
 ### Changed