+++ CHANGELOG.md
@@ -37,8 +37,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - The `tango run` command will use the lightweight base executor (single process) by default.
     To use the multi-process executor, set `-j/--parallelism` to 1 or higher or -1 to use all available CPU cores.
 
+## [v0.10.1](https://github.com/allenai/tango/releases/tag/v0.10.1) - 2022-07-26
+
 ### Fixed
 
+- Fixed issue where the StepInfo config argument could be parsed into a Step. 
 - Restored capability to run tests out-of-tree.
 
 ## [v0.10.0](https://github.com/allenai/tango/releases/tag/v0.10.0) - 2022-07-07